Add all current ebuild patches, for packages:   *  app-emulation/wine   *  app-emulation/wine-staging   *  app-emulation/wine-vanilla to gentoo-wine-ebuild-common repository.
**UPDATED: GSTREAMER 1.0 PATCH**
The Gentoo GStreamer 1.0 patch, for Wine 1.8, is re-segmented into the
original patchset, consisting of patches 01-11 (11 patches total). These
patches can then therefore be selectively applied at a Git Commit level.

**NEW: FREETYPE 2.8.1 PATCHES**
Add 3 patch freetype 2.8.1 patchset to repository.
The patch:
  *  wine-2.18-freetype-2.8.1-implmenet_minimum_em_size_required_by_opentype_1.8.2.patch
is a binary patch.

**NEW: PATCH COMMIT FILTERING**
All patches that were commited to the Wine Stable 1.8.x branch now have the
Git Commit hash from this (stable) branch added to the header of the patch.
This allows the live 9999 versions of:
  *  app-emulation/wine
  *  app-emulation/wine-staging
  *  app-emulation/wine-vanilla
to filter patch Git Commits, that are already commited as a parent of a
specific Git Commit, in the Wine Git tree.